# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API BB Serviços 1.0.0",
        "description": "A documentação BB Serviços, tem como finalidade facilitar a utilização dos endpoints para consumo de APIs externas.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/pedido/{Pedido}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Status do pedido",
                "description": "Retorna o status atual do pedido",
                "operationId": "a43a6a0888b2085e05f06d12cb67a3c8",
                "parameters": [
                    {
                        "name": "Pedido",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/produtos/{produto}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Produtos em estoque",
                "description": "Retorna o produto armazenado em nossos estoques",
                "operationId": "bba1217799c6fa59406a232707b7b0f8",
                "parameters": [
                    {
                        "name": "produto",
                        "in": "path",
                        "description": "Nome do produto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "produto": {
                                            "description": "Nome do produto",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/produtos": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar todos os produtos",
                "description": "Retorna todos os produtos armazenados em nossos estoques",
                "operationId": "fb7e33cc3ffe3f09dcd888dd0a968052",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "produto": {
                                                "description": "Nome do produto",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}